Adaptation of drupalextension_remote functionality into parent.
The functionality look to provide base level integration / driver for the
drupalextension_remote module which allows behat to interactively work
with a remote drupal instance using regimented api.

The configuration allows for a different api client to be used, ultimately
providing the ability to extend the API further should a different test
regimen need to do so in the future.

TODO: Documentation updates.